! some scruit items

+++ sources/ElkArte/AdminController/Maintenance.php
@@ -230,7 +230,7 @@ class Maintenance extends \ElkArte\AbstractController
 		$plain_limit = 240000;
 
 		// Last thing: are we able to gain time?
-		$current_time_limit = ini_get('max_execution_time');
+		$current_time_limit = (int) ini_get('max_execution_time');
 		@set_time_limit(159); //something strange just to be sure
 		$new_time_limit = ini_get('max_execution_time');
 		@set_time_limit($current_time_limit);
